Shift version history to separate file
